backend/requirements.txt
aiohttp==3.6.2
aniso8601==7.0.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.3
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
Django==3.1.1
django-channels-graphql-ws==0.7.4
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.12.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
multidict==4.7.6
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
yarl==1.6.0
zope.interface==5.1.0
